@@ -22,18 +22,18 @@
     "framer-motion": "^6.0.0",
     "hamburger-react": "^2.4.1",
     "immer": "^9.0.12",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-double-marquee": "^1.1.0",
     "react-icons": "^4.3.1",
     "react-intersection-observer": "^8.33.1",
+    "react-responsive": "^9.0.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "react-select": "^5.7.0",
     "react-spinners": "^0.11.0",
     "react-use-mounted": "^2.0.0",
-    "simple-zustand-devtools": "^1.1.0",
     "web-vitals": "^2.1.0",
     "zustand": "^3.6.9"
   },
@@ -60,5 +60,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "simple-zustand-devtools": "^1.1.0"
   }
 }
