@@ -1,19 +1,22 @@
+import "./index.css";
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
+import { createRoot } from 'react-dom/client';
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
 import { PrismicProvider } from "@prismicio/react";
 import { client } from "./prismic";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <PrismicProvider client={client}>
       <App />
     </PrismicProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
