@@ -1,32 +1,25 @@
 import styled from "@emotion/styled";
-import * as prismic from "@prismicio/client";
-import { usePrismicDocumentsByType } from "@prismicio/react";
 import React from "react";
-import FadeIn from "../../Animations/FadeIn";
-import Divider from "../Divider";
 import BroadcastItem from "./BroadcastItem";
 
 const Container = styled.section`
-  h3 {
-    text-transform: capitalize;
+  .list {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
   }
 `
 
-const BroadcastList = ({ tag }) => {
-  const [documents] = usePrismicDocumentsByType("broadcasts", {
-    predicates: [prismic.predicate.at("document.tags", [tag])],
-      fetchLinks: "data.hostedby.title",
-  });
-  if (!documents || documents.results.length < 1) return <></>;
+const BroadcastList = ({ broadcasts }) => {
+  if (!broadcasts) return <></>;
   return (
     <Container>
-      <FadeIn>
-        <h3>{tag}</h3>
-      </FadeIn>
-      {documents.results.map((broadcast) => (
-        <BroadcastItem broadcast={broadcast} key={broadcast.id} />
-      ))}
-      <Divider />
+
+      <div class="list">
+
+        {broadcasts.map((broadcast) => (
+          <BroadcastItem broadcast={broadcast} key={broadcast.id} />
+        ))}
+      </div>
     </Container>
   );
 };
