@@ -8,7 +8,7 @@ import Layout from "../Components/Layout";
 import NotFound from "../Components/NotFound";
 import PageLoader from "../Components/PageLoader";
 import Tags from "../Components/Tags";
-import TeaserImage from "../Components/TeaserImage/TeaserImage";
+import ThumbnailImage from "../Components/TeaserImage/ThumbnailImage";
 import useThemeStore from "../Stores/ThemeStore";
 import { SoundcloudPlayer } from "./Show";
 
@@ -41,7 +41,7 @@ const Broadcast = () => {
         <Container>
           <FadeIn>
             <Header>
-              <TeaserImage image={document[0].data.image} />
+              <ThumbnailImage image={document[0].data.image} />
             </Header>
           </FadeIn>
           <FadeIn>
