@@ -1,6 +1,6 @@
 import produce from "immer";
+import { mountStoreDevtool } from "simple-zustand-devtools";
 import create from "zustand";
-import { mountStoreDevtool } from 'simple-zustand-devtools';
 
 const useBroadcastStore = create(
   produce((set) => ({
