@@ -1,7 +1,7 @@
 import React from "react";
 
 const Logo = () => {
-    return (<svg width="200" height="64" viewBox="0 0 200 64" fill="none" xmlns="http://www.w3.org/2000/svg">
+    return (<svg width="153" height="48" viewBox="0 0 200 64" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path d="M35.4003 38.5434L37.1089 36.8684C36.4381 37.3131 35.8582 37.8816 35.4003 38.5434V38.5434Z" fill="currentColor" />
         <path d="M64.0235 63.7845H60.6365L36.6985 39.8549V63.7845H34.268V42.1916C34.2637 40.8914 34.6564 39.6209 35.3937 38.55V38.55L37.1022 36.875C38.1534 36.1785 39.3873 35.809 40.6483 35.813H61.4372V38.2452H38.4791L64.0235 63.7845Z" fill="currentColor" />
         <path d="M104.298 63.7896H74.8962C74.0566 63.7925 73.2248 63.6292 72.4486 63.3093C71.6724 62.9893 70.9672 62.519 70.3735 61.9253C69.7799 61.3317 69.3095 60.6264 68.9896 59.8502C68.6696 59.074 68.5064 58.2422 68.5093 57.4027V54.9705C68.5062 54.1308 68.6692 53.2988 68.9891 52.5224C69.3089 51.746 69.7793 51.0406 70.3729 50.4468C70.9666 49.8529 71.6719 49.3824 72.4482 49.0624C73.2245 48.7423 74.0565 48.579 74.8962 48.5819H101.864V38.2419H69.8275V35.808H97.9113C98.7508 35.8051 99.5826 35.9684 100.359 36.2883C101.135 36.6083 101.84 37.0786 102.434 37.6723C103.028 38.2659 103.498 38.9712 103.818 39.7474C104.138 40.5236 104.301 41.3554 104.298 42.195V63.7896ZM101.864 61.3558V51.0157H70.9431V61.3558H101.864Z" fill="currentColor" />
